+++ lib/experiments-page.php
@@ -86,7 +86,7 @@ function gutenberg_initialize_experiments_settings() {
 		'gutenberg-experiments',
 		'gutenberg_experiments_section',
 		array(
-			'label' => __( 'Disables the TinyMCE and Classic block', 'gutenberg' ),
+			'label' => __( 'Disables the TinyMCE and Classic block.', 'gutenberg' ),
 			'id'    => 'gutenberg-no-tinymce',
 		)
 	);
@@ -172,18 +172,6 @@ function gutenberg_initialize_experiments_settings() {
 		array(
 			'label' => __( 'Enables access to a Quick Edit panel in the Site Editor Pages experience.', 'gutenberg' ),
 			'id'    => 'gutenberg-quick-edit-dataviews',
-		)
-	);
-
-	add_settings_field(
-		'gutenberg-full-page-client-side-navigation',
-		__( 'iAPI: full page client side navigation', 'gutenberg' ),
-		'gutenberg_display_experiment_field',
-		'gutenberg-experiments',
-		'gutenberg_experiments_section',
-		array(
-			'label' => __( 'Enables full-page client-side navigation with the Interactivity API, updating HTML while preserving application state.', 'gutenberg' ),
-			'id'    => 'gutenberg-full-page-client-side-navigation',
 		)
 	);